Refactor cli functions out of main to be usable in test suite
